% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple.integer}
\alias{triple.integer}
\title{triple.integer}
\usage{
\method{triple}{integer}(x, ...)
}
\arguments{
\item{...}{}

\item{number}{a vector of integers}
}
\value{
A vector the same length as \code{number} containing the returning values
}
\description{
compute the triple of an integer
}
\examples{
triple.integer(2)
triple.integer(1:4)
}
