% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple.numeric}
\alias{triple.numeric}
\title{triple.numeric}
\usage{
\method{triple}{numeric}(x, ...)
}
\arguments{
\item{...}{further arguments passed to the generic method.}

\item{number}{a numeric vector}
}
\value{
A vector the same length as \code{number} containing the returning values
}
\description{
compute the triple of a real number
}
\examples{
triple.numeric(2)
triple.numeric(1:4)
}
