% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triple.R
\name{triple}
\alias{triple}
\title{triple}
\usage{
triple(x, ...)
}
\arguments{
\item{x}{a numeric vector}

\item{...}{}
}
\value{
A vector the same length as \code{x} containing the returning values
}
\description{
triple
}
\examples{
triple(2)
triple(1:4)
}
